app/requirements.txt
# Core
streamlit>=1.28.0
python-dotenv>=1.0.0

# LLM Inference
llama-cpp-python>=0.2.0
huggingface-hub>=0.19.0

# Data & APIs
requests>=2.31.0
urllib3>=2.0.0

# Visualization
plotly>=5.18.0
pandas>=2.0.0

# Utilities
psutil>=5.9.0